ai_agents/requirements.txt
openai-agents
requests
python-dotenv
pydantic
orjson
fastapi
uvicorn
httpx
astroquery
pandas
scikit-learn
annotated-types
anyio